LayoutTests/inspector/console/console-format-style-whitelist-expected.txt

CONSOLE MESSAGE: line 8: %cColors are awesome.
CONSOLE MESSAGE: line 9: %cSo are fonts!
CONSOLE MESSAGE: line 10: %cAnd borders and margins and paddings!
CONSOLE MESSAGE: line 11: %ctext-* is fine by us!
CONSOLE MESSAGE: line 13: %cDisplay, on the other hand, is bad news.
CONSOLE MESSAGE: line 14: %cAnd position too.
Tests that console logging dumps properly styled messages, and that the whole message gets the same style, regardless of multiple %c settings.

Colors are awesome. console-format-style-whitelist.html:8
Styled text #0: color: blue;
So are fonts! console-format-style-whitelist.html:9
Styled text #0: font-style: normal; font-variant: normal; font-weight: normal; font-size: 1em; line-height: normal; font-family: Helvetica;
And borders and margins and paddings! console-format-style-whitelist.html:10
Styled text #0: border: 1px solid red; margin: 20px; padding: 10px;
text-* is fine by us! console-format-style-whitelist.html:11
Styled text #0: text-decoration: none;
Display, on the other hand, is bad news. console-format-style-whitelist.html:13
Styled text #0: NO STYLES DEFINED
And position too. console-format-style-whitelist.html:14
Styled text #0: NO STYLES DEFINED
